+++ nav2_costmap_2d/include/nav2_costmap_2d/denoise/image_processing.hpp
